+++ napari/_qt/layers/__init__.py
@@ -1,1 +1,1 @@
-from .util import create_qt_controls
+from .utils import create_qt_controls
