+++ napari/_qt/layers/utils.py
