+++ napari/_tests/test_key_bindings.py
@@ -3,8 +3,8 @@ from unittest.mock import Mock
 from vispy import keys
 
 
-def test_viewer_keybindings(viewer_factory):
-    """Test adding keybindings to the viewer
+def test_viewer_key_bindings(viewer_factory):
+    """Test adding key bindings to the viewer
     """
     np.random.seed(0)
     view, viewer = viewer_factory()
@@ -71,8 +71,8 @@ def test_viewer_keybindings(viewer_factory):
     mock_shift_release.reset_mock()
 
 
-def test_layer_keybindings(viewer_factory):
-    """Test adding keybindings to a layer
+def test_layer_key_bindings(viewer_factory):
+    """Test adding key bindings to a layer
     """
     np.random.seed(0)
     view, viewer = viewer_factory()
