+++ napari/layers/base/_base_constants.py
