+++ napari/layers/image/_image_constants.py
