+++ napari/layers/image/_image_utils.py
