+++ napari/layers/utils/_tests/test_layer_utils.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 from dask import array as da
 
-from ..layer_utils import (
+from napari.layers.utils.layer_utils import (
     calc_data_range,
     increment_unnamed_colormap,
     segment_normal,
