+++ napari/layers/utils/layer_utils.py
